3 days of various tests about IPC / RPC
I ended up with xml-rpc because I really need rpc and It is the only solution that I could get working. Wamp would have been better with the RPC + IPC support and it's bidirectional, realtime communication but sadly (Py)Qt5 at the time of writing only support websockets.

This commit is a backup commit