Found a way to hide and show widgets deleted a useless tutorial (it was for pyqt4)
Little explanation for the previous commit:
I splited the main gui file in 3 separated files, one for each tab and one for the main windows.
I deleted all the useless dc terms and added "durée"
I generated a new test database to reflect the data-structure changes
The search function now work as expected
Still need to format and display the answer of the search query.