It's now possible to losslessly digitise a VHS using ffv1 and flac
Next step is to use the blackmagic card directly and not the current VHS player -> scart_to_hdmi_converter -> blackmagic card setup.
This has to be done because the scart to hdmi converter use a lot of postprocessing to make the video look decent and we want to alter the video as minimally as possible in a lossless capture.